+++ packages/did-ui-react/src/components/SignIn/index.tsx
@@ -271,6 +271,7 @@ const SignIn = forwardRef(
             verifierId: res.verifier.id,
             verificationDoc: res.verificationDoc,
             signature: res.signature,
+            zkLoginInfo: res.zkLoginInfo,
           },
         ];
         setApprovedList(list);
@@ -300,6 +301,9 @@ const SignIn = forwardRef(
             const result = await verifySocialToken({
               accountType,
               token: authenticationInfo?.authToken,
+              idToken: authenticationInfo?.idToken,
+              nonce: authenticationInfo?.nonce,
+              timestamp: authenticationInfo?.timestamp,
               guardianIdentifier: identifier,
               verifier,
               chainId,
@@ -309,12 +313,15 @@ const SignIn = forwardRef(
             });
             setLoading(false);
 
-            if (!result?.signature || !result?.verificationDoc) throw 'Verify social login error';
+            if (!result?.zkLoginInfo && (!result?.signature || !result?.verificationDoc)) {
+              throw 'Verify social login error';
+            }
             onStep2OfSignUpFinish(
               {
                 verifier,
                 verificationDoc: result.verificationDoc,
                 signature: result.signature,
+                zkLoginInfo: result.zkLoginInfo,
               },
               value,
             );