+++ packages/services/src/types/verification.ts
@@ -37,8 +37,13 @@ export type VerifyVerificationCodeParams = {
 };
 
 export type VerifyVerificationCodeResult = {
-  verificationDoc: string;
-  signature: string;
+  verificationDoc?: string;
+  signature?: string;
+  zkLoginInfo?: ZKLoginInfo;
+};
+
+export type VerifyZKLoginResult = {
+  guardianIdentifierHash: string;
 };
 
 export type SendAppleUserExtraInfoParams = {
@@ -68,6 +73,38 @@ export interface VerifierSocialTokenParams extends BaseVerifyTokenParams {
 export interface VerifyAppleTokenParams extends BaseVerifyTokenParams {
   identityToken: string;
   caHash?: string;
+}
+
+export type VerifyZKPortkeyParams = {
+  type: string;
+  accessToken?: string;
+  jwt?: string;
+  verifierId?: string;
+  chainId: ChainId;
+  operationType: OperationTypeEnum;
+};
+
+export type VerifyZKLoginParams = {
+  verifyToken: VerifyZKPortkeyParams;
+  jwt?: string;
+  salt: string;
+  kid: string;
+  nonce?: string;
+  timestamp: number;
+  managerAddress: string;
+};
+
+export interface ZKLoginInfo {
+  identifierHash: string;
+  poseidonIdentifierHash: string;
+  identifierHashType: number;
+  salt: string;
+  zkProof: string;
+  jwt: string;
+  nonce: string;
+  timestamp: number;
+  managerAddress: string;
+  circuitId: string;
 }
 
 export type SendAppleUserExtraInfoResult = { userId: string };
@@ -114,6 +151,7 @@ export interface IVerificationService {
   getAppleUserExtraInfo(params: GetAppleUserExtraInfoParams): Promise<getAppleUserExtraInfoResult>;
   verifyGoogleToken(params: VerifierSocialTokenParams): Promise<VerifyVerificationCodeResult>;
   verifyAppleToken(params: VerifyAppleTokenParams): Promise<VerifyVerificationCodeResult>;
+  verifyZKLogin(params: any): Promise<VerifyZKLoginResult>;
   verifyTelegramToken(params: VerifierSocialTokenParams): Promise<VerifyVerificationCodeResult>;
   verifyTwitterToken(
     params: VerifierSocialTokenParams,